// apps/web/src/types/vitrine.ts
export type Vitrine = {
  id: string;
  title: string;
  description: string | null;
  vimeoShowcaseId: string | null;
  vimeoSource: "MANUAL" | "VIMEO_SHOWCASE";
  status?: "ACTIVE" | "EDITING" | "INACTIVE";
  createdAt: string;
  updatedAt?: string;
  account?: { id: string; name: string };
  videoCount?: number;
};